temarlije/backend/requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
readability-lxml
python-dotenv
openai>=1.0.0

